+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/send/media/SendAudio.kt
@@ -88,7 +88,7 @@ data class SendAudioData internal constructor(
     @SerialName(replyMarkupField)
     @Optional
     override val replyMarkup: KeyboardMarkup? = null
-) : Data<RawMessage>,
+) : DataRequest<RawMessage>,
     SendMessageRequest<RawMessage>,
     ReplyingMarkupSendMessageRequest<RawMessage>,
     TextableSendMessageRequest<RawMessage>,