+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/send/media/SendDocument.kt
@@ -72,7 +72,7 @@ data class SendDocumentData internal constructor(
     @SerialName(replyMarkupField)
     @Optional
     override val replyMarkup: KeyboardMarkup? = null
-) : Data<RawMessage>,
+) : DataRequest<RawMessage>,
     SendMessageRequest<RawMessage>,
     ReplyingMarkupSendMessageRequest<RawMessage>,
     TextableSendMessageRequest<RawMessage>,