+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/send/media/base/Data.kt
@@ -1,5 +1,10 @@
 package com.github.insanusmokrassar.TelegramBotAPI.requests.send.media.base
 
-import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.SimpleRequest
-
-interface Data<T: Any> : SimpleRequest<T>
+@Deprecated(
+    "Renamed to DataRequest",
+    ReplaceWith(
+        "DataRequest",
+        "com.github.insanusmokrassar.TelegramBotAPI.requests.send.media.base.DataRequest"
+    )
+)
+typealias Data<T> = DataRequest<T>